remove unneccsery repo and data source
