+++ lib/core/common/class/firestore_services.dart
@@ -2,7 +2,7 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:injectable/injectable.dart';
 import 'package:kamn/features/sports_service_providers/data/model/playground_request_model.dart';
 
-@singleton
+@lazySingleton
 class FirestoreService {
   late FirebaseFirestore _firestore;
