+++ lib/features/sports_service_providers/data/repository/service_providers_repository.dart
@@ -1,10 +1,10 @@
 import 'dart:io';
 
-import 'package:connectivity_plus/connectivity_plus.dart';
 import 'package:fpdart/fpdart.dart';
 import 'package:injectable/injectable.dart';
 import 'package:kamn/core/erorr/faliure.dart';
 import 'package:kamn/core/utils/try_and_catch.dart';
+import 'package:kamn/features/sports/data/models/playground_model.dart';
 import 'package:kamn/features/sports_service_providers/data/data_source/service_providers_remote_data_source.dart';
 import 'package:kamn/features/sports_service_providers/data/model/playground_request_model.dart';
 import 'dart:async';
@@ -14,6 +14,11 @@ abstract class ServiceProvidersRepository {
       PlaygroundRequestModel playground);
   Future<Either<Faliure, List<String>>> addFImagesToStorage(List<File> images);
   Future<Either<Faliure, bool>> deleteImagesFromStorage(List<File> images);
+  Future<Either<Faliure, List<PlaygroundRequestModel>>>
+      getPlaygroundsRequests();
+  Future<Either<Faliure, void>> addToFirebase(PlaygroundModel playground);
+  Future<Either<Faliure, void>> updateState(
+      PlaygroundRequestModel playground, Map<String, dynamic> data);
 }
 
 @Injectable(as: ServiceProvidersRepository)
@@ -42,6 +47,32 @@ class ServiceProvidersRepositoryImpl implements ServiceProvidersRepository {
       List<File> images) async {
     return executeTryAndCatchForRepository(() async {
       return await dataSource.deleteImagesFromStorage(images);
+    });
+  }
+
+  @override
+  Future<Either<Faliure, List<PlaygroundRequestModel>>>
+      getPlaygroundsRequests() {
+    return executeTryAndCatchForRepository(() async {
+      var data = await dataSource.getPlaygroundsRequests();
+      return data.map((value) {
+        return PlaygroundRequestModel.fromMap(value);
+      }).toList();
+    });
+  }
+
+  @override
+  Future<Either<Faliure, void>> addToFirebase(PlaygroundModel playground) {
+    return executeTryAndCatchForRepository(() async {
+      return await dataSource.addToFiresore(playground);
+    });
+  }
+
+  @override
+  Future<Either<Faliure, void>> updateState(
+      PlaygroundRequestModel playground, Map<String, dynamic> data) {
+    return executeTryAndCatchForRepository(() async {
+      return await dataSource.updateState(playground, data);
     });
   }
 }
