+++ lib/features/sports_service_providers/presentation/cubit/service_provider_ground_details/service_provider_ground_details_cubit.dart
@@ -3,7 +3,7 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:injectable/injectable.dart';
 import 'package:kamn/features/sports/data/models/playground_model.dart';
 import 'package:kamn/features/sports_service_providers/data/model/playground_request_model.dart';
-import 'package:kamn/features/sports_service_providers/data/repository/service_provider_ground_details_repository.dart';
+import 'package:kamn/features/sports_service_providers/data/repository/service_providers_repository.dart';
 import 'package:kamn/features/sports_service_providers/presentation/cubit/service_provider_ground_details/service_provider_ground_details_state.dart';
 
 @injectable
@@ -13,7 +13,7 @@ class ServiceProviderGroundDetailsCubit
       : super(ServiceProviderGroundDetailsState(
             state: ServiceProviderGroundDetailsStatus.initial));
 
-  ServiceProviderGroundDetailsRepository repository;
+  ServiceProvidersRepository repository;
 
   TextEditingController commentController = TextEditingController();
   Future<void> accept(PlaygroundModel playground) async {
